Extract more constants
