+++ tests/block_creation_time_test.py
@@ -2,7 +2,7 @@ import unittest
 import pytest
 
 from substrateinterface import SubstrateInterface
-from tools.utils import WS_URL
+from tools.constants import WS_URL
 from peaq.utils import wait_for_n_blocks, get_block_height
 from tools.block_creation_utils import get_block_creation_times
