+++ tests/bridge_batch_test.py
@@ -1,7 +1,7 @@
 import unittest
 import pytest
 
-from tools.utils import WS_URL, ETH_URL
+from tools.constants import WS_URL, ETH_URL
 # from tools.runtime_upgrade import wait_until_block_height
 from tools.peaq_eth_utils import get_contract
 from tools.peaq_eth_utils import GAS_LIMIT, get_eth_info
@@ -9,7 +9,7 @@ from tools.peaq_eth_utils import get_eth_chain_id
 from substrateinterface import SubstrateInterface
 from peaq.utils import ExtrinsicBatch
 from web3 import Web3
-from tools.utils import KP_GLOBAL_SUDO
+from tools.constants import KP_GLOBAL_SUDO
 from tools.peaq_eth_utils import sign_and_submit_evm_transaction
 from tools.peaq_eth_utils import generate_random_hex
