+++ tests/bridge_parachain_staking_test.py
@@ -3,7 +3,7 @@ import unittest
 from tests.utils_func import restart_parachain_and_runtime_upgrade
 from tools.runtime_upgrade import wait_until_block_height
 from substrateinterface import SubstrateInterface
-from tools.utils import WS_URL, ETH_URL
+from tools.constants import WS_URL, ETH_URL
 from tools.peaq_eth_utils import sign_and_submit_evm_transaction
 from peaq.utils import ExtrinsicBatch
 from tools.peaq_eth_utils import get_contract
@@ -11,7 +11,7 @@ from tools.peaq_eth_utils import get_eth_chain_id
 # , calculate_evm_default_addr
 from tools.peaq_eth_utils import GAS_LIMIT, get_eth_info
 from tools.evm_claim_sign import calculate_claim_signature, claim_account
-from tools.utils import KP_GLOBAL_SUDO, KP_COLLATOR
+from tools.constants import KP_GLOBAL_SUDO, KP_COLLATOR
 from peaq.utils import get_block_hash
 from web3 import Web3
