+++ tests/bridge_rbac_test.py
@@ -2,7 +2,7 @@ import pytest
 
 from substrateinterface import SubstrateInterface, Keypair, KeypairType
 from tools.peaq_eth_utils import sign_and_submit_evm_transaction
-from tools.utils import WS_URL, ETH_URL
+from tools.constants import WS_URL, ETH_URL
 from peaq.eth import calculate_evm_account, calculate_evm_addr
 from tools.peaq_eth_utils import get_eth_chain_id
 from tools.peaq_eth_utils import call_eth_transfer_a_lot, get_contract, generate_random_hex, TX_SUCCESS_STATUS
