+++ tests/pallet_did_test.py
@@ -4,7 +4,7 @@ import pytest
 
 from substrateinterface import SubstrateInterface, Keypair
 from tools.utils import get_balance_reserve_value
-from tools.utils import WS_URL
+from tools.constants import WS_URL
 from peaq.utils import ExtrinsicBatch
 from peaq.did import did_add_payload, did_update_payload, did_remove_payload, did_rpc_read
