+++ tests/upgrade_test.py
@@ -2,7 +2,7 @@ import sys
 sys.path.append('./')
 
 from substrateinterface import SubstrateInterface, Keypair, KeypairType
-from tools.utils import WS_URL
+from tools.constants import WS_URL
 from peaq.utils import show_extrinsic
 # from scalecodec.base import RuntimeConfiguration
 # from scalecodec.base import ScaleBytes
