+++ tests/xcm_transfer_test.py
@@ -6,12 +6,12 @@ import unittest
 from tests.utils_func import restart_parachain_and_runtime_upgrade
 from tools.runtime_upgrade import wait_until_block_height
 from substrateinterface import SubstrateInterface, Keypair
-from tools.utils import WS_URL, RELAYCHAIN_WS_URL, ACA_WS_URL, PARACHAIN_WS_URL
-from tools.utils import get_modified_chain_spec
+from tools.constants import WS_URL, RELAYCHAIN_WS_URL, ACA_WS_URL, PARACHAIN_WS_URL
+from tools.utils import get_modified_chain_spec, batch_fund
 from peaq.utils import get_account_balance
 from peaq.utils import ExtrinsicBatch
 from peaq.sudo_extrinsic import fund
-from tools.utils import KP_GLOBAL_SUDO, ACA_PD_CHAIN_ID, batch_fund
+from tools.constants import KP_GLOBAL_SUDO, ACA_PD_CHAIN_ID
 from tools.asset import setup_asset_if_not_exist
 from tools.asset import batch_register_location, batch_set_units_per_second, setup_xc_register_if_not_exist
 from tools.asset import setup_aca_asset_if_not_exist
