+++ tools/get_eth_block.py
@@ -1,5 +1,5 @@
 from web3 import Web3
-from tools.utils import ETH_URL
+from tools.constants import ETH_URL
 
 w3 = Web3(Web3.HTTPProvider(ETH_URL))
 block = w3.eth.get_block('latest')
