+++ tools/measure_websocket.py
@@ -1,7 +1,7 @@
 from substrateinterface import SubstrateInterface
 import datetime
 import time
-from tools.constants import BLOCK_GENERATIME_TIME
+from tools.constants import BLOCK_GENERATE_TIME
 
 
 WS = 'wss://docker-test.peaq.network'
@@ -14,5 +14,5 @@ while 1:
     substrate.get_block_metadata()
     after = datetime.datetime.now()
     print(f'{after}: {after - before}')
-    time.sleep(BLOCK_GENERATIME_TIME)
+    time.sleep(BLOCK_GENERATE_TIME)
     substrate.close()
