+++ tools/monkey/monkey_3rd_substrate_interface.py
@@ -1,7 +1,7 @@
 from substrateinterface.exceptions import SubstrateRequestException, ExtrinsicNotFound
 from scalecodec.types import GenericExtrinsic
 from substrateinterface.base import ExtrinsicReceipt
-from tools.constants import BLOCK_GENERATIME_TIME
+from tools.constants import BLOCK_GENERATE_TIME
 import time
 
 
@@ -11,7 +11,7 @@ def _wait_finalization(substrate, included_block):
         print(f'Checking finalized block {finalized_block} and included block {included_block}')
         if finalized_block >= included_block:
             break
-        time.sleep(BLOCK_GENERATIME_TIME)
+        time.sleep(BLOCK_GENERATE_TIME)
 
 
 def monkey_submit_extrinsic(self, extrinsic: GenericExtrinsic, wait_for_inclusion: bool = False,
@@ -28,7 +28,7 @@ def monkey_submit_extrinsic(self, extrinsic: GenericExtrinsic, wait_for_inclusio
     if not wait_for_inclusion and not wait_for_finalization:
         return result
 
-    time.sleep(BLOCK_GENERATIME_TIME * 3)
+    time.sleep(BLOCK_GENERATE_TIME * 3)
     now_block_num = self.get_block_number(None)
     included_block = None
     tx_identifier = None
