+++ tools/monkey/monkey_reorg_batch.py
@@ -2,7 +2,7 @@ from peaq.utils import ExtrinsicBatch
 from substrateinterface.exceptions import SubstrateRequestException
 import time
 from peaq.utils import show_extrinsic
-from tools.constants import BLOCK_GENERATIME_TIME
+from tools.constants import BLOCK_GENERATE_TIME
 
 BACKTRACE_BLOCK_NUM = 10
 
@@ -37,7 +37,7 @@ def monkey_execute_extrinsic_batch(self, substrate, kp_src, batch,
             if 'invalid' in str(e):
                 print(f'Error: {e}, {batch}')
                 print('Wait for 30 seconds')
-                time.sleep(BLOCK_GENERATIME_TIME * 5)
+                time.sleep(BLOCK_GENERATE_TIME * 5)
                 block_hash = _backtrace_blocks_by_extrinsic(
                     substrate, self.submit_extrinsic.extrinsic_hash.hex())
                 if block_hash:
