+++ tools/restart.py
@@ -8,8 +8,8 @@ sys.path.append('.')
 import time
 from python_on_whales import docker, DockerClient
 from substrateinterface import SubstrateInterface
-from tools.utils import WS_URL
-from tools.constants import BLOCK_GENERATIME_TIME
+from tools.constants import WS_URL
+from tools.constants import BLOCK_GENERATE_TIME
 from websocket import WebSocketConnectionClosedException
 
 
@@ -35,7 +35,7 @@ def restart_parachain_launch():
         except (ConnectionResetError, WebSocketConnectionClosedException) as e:
             print(f'Cannot connect to {WS_URL}, {e}')
             count_down += 5
-            time.sleep(BLOCK_GENERATIME_TIME)
+            time.sleep(BLOCK_GENERATE_TIME)
             continue
         except Exception:
             raise IOError(f'Cannot connect to {WS_URL}')
